feat: dateInputV2
